Algorithms/0451.sort-characters-by-frequency/SortCharactersByFrequency.java
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

class SortCharactersByFrequency {
    public String frequencySort(String s) {
        Map<Character, Integer> freq = new HashMap<>();
        for (char c : s.toCharArray())
            freq.put(c, freq.getOrDefault(c, 0) + 1);

        List<Character> list = new ArrayList<>(freq.keySet());
        list.sort((a, b) -> freq.get(b) - freq.get(a));
        StringBuilder sb = new StringBuilder();
        for (char c : list) {
            int frequency = freq.get(c);
            sb.append(String.valueOf(c).repeat(frequency));
        }
        return sb.toString();
    }
}